refactor: 便利系コマンドを一つのクラスにまとめる
